Add averagePnL, net ROI and updated data importing logic
In the TradeLog module, I added a new property 'averagePnL' and updated the logic for generating trade response to include 'averagePnL'. Renamed the function 'getTradeStartdate' to 'getTradeStartDate'. Introduced function 'symbolMargin' to calculate the margin for specific symbols. Also, updated return and display logic for 'net roi' and 'net p&l'.

In the Cost entity, 'tradeFillRefs' property was added which serves as the reference for fills.

In the ImporterService, refined the logic for parsing CSV data and improved the cost handling mechanism. The older simplistic approach, replaced by a more robust and accurate mechanism, included costs divided over multiple trades. It now efficiently handles costs by the time of trade.

Lastly, corrected some typos in the README file and updated the TODO list.